# docs/phoenix-openapi.yaml
openapi: 3.0.3
info:
  title: Phoenix ERP Mock API
  version: "1.0.0"
  description: |
    REST API for the techbold **AI Sphinx** track. Your app uses this
    mock ERP to load assigned tickets, read the affected customer system, set ticket
    status, and write back an activity log.

    **Auth:** every endpoint requires your team's bearer token:
    `Authorization: Bearer <PHOENIX_API_TOKEN>`

    **Workflow:** `GET /me/tickets` → `GET /tickets/{id}/customer-system` (SSH details)
    → fix the VM over SSH → `POST /activities/create` → `PATCH /tickets/{id}/status`.

    Notes:
    - Ticket descriptions are the customer's report (symptom only).
    - IDs are 64-bit integers.
    - The private SSH key is **not** returned by the API; you receive the `.pem` separately.

servers:
  - url: "{baseUrl}"
    description: Phoenix mock base URL (provided on Builder Base)
    variables:
      baseUrl:
        default: http://68.210.101.85:8000
        enum:
          - http://68.210.101.85:8000
          - http://host.docker.internal:8000
          - http://localhost:8000

security:
  - bearerAuth: []

tags:
  - name: Phoenix
    description: Hacker-facing ERP endpoints

paths:
  /api/v1/me:
    get:
      tags: [Phoenix]
      summary: Get the logged-in technician
      operationId: getMe
      responses:
        "200":
          description: The technician identity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /api/v1/me/tickets:
    get:
      tags: [Phoenix]
      summary: List the technician's assigned tickets
      description: Returns the tickets for the team's current mode. Optional filtering and sorting.
      operationId: listMyTickets
      parameters:
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/TicketStatus" }
        - name: priority
          in: query
          required: false
          schema: { type: string, example: high }
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [date, priority, status]
            default: date
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Ticket" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /api/v1/tickets/{ticket_id}:
    get:
      tags: [Phoenix]
      summary: Get a single ticket
      operationId: getTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: The ticket
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  /api/v1/tickets/{ticket_id}/customer-system:
    get:
      tags: [Phoenix]
      summary: Get the customer system (SSH target) for a ticket
      operationId: getCustomerSystem
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: The customer system info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerSystem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  /api/v1/tickets/{ticket_id}/status:
    patch:
      tags: [Phoenix]
      summary: Set a ticket's status
      operationId: setTicketStatus
      parameters:
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StatusUpdate" }
      responses:
        "200":
          description: The updated ticket
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }

  /api/v1/customers/{customer_id}:
    get:
      tags: [Phoenix]
      summary: Get customer information (incl. system)
      operationId: getCustomer
      parameters:
        - name: customer_id
          in: path
          required: true
          schema: { type: integer, format: int64, example: 5001 }
      responses:
        "200":
          description: The customer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  /api/v1/activities/create:
    post:
      tags: [Phoenix]
      summary: Create an activity (write documentation back to the ERP)
      description: |
        The team and technician are derived from your token. `ticket_id` must be one of
        your currently-assigned tickets. Fill the documentation fields properly — they are graded.
      operationId: createActivity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActivityCreate" }
      responses:
        "201":
          description: The created activity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activity" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }

  /api/v1/me/reset:
    post:
      tags: [Phoenix]
      summary: Reset your activities and reboot your VMs
      description: Clears all of your team's activities and requests a reboot of your VMs.
      operationId: resetMe
      responses:
        "200":
          description: Reset result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleMessage" }
        "401": { $ref: "#/components/responses/Unauthorized" }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Your team's static Phoenix API token.

  parameters:
    TicketId:
      name: ticket_id
      in: path
      required: true
      schema: { type: integer, format: int64, example: 7001 }

  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found (or not assigned in the current mode)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request failed validation (FastAPI returns `detail` as an array of errors)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/HTTPValidationError" }

  schemas:
    TicketStatus:
      type: string
      enum: [OPEN, PENDING, DONE]

    Employee:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1001 }
        firstname: { type: string, example: Max }
        lastname: { type: string, example: Mustermann }
        username: { type: string, example: m.mustermann }
        teamname: { type: string, example: Remote Support }
      required: [id, firstname, lastname, username, teamname]

    Ticket:
      type: object
      properties:
        id: { type: integer, format: int64, example: 7001 }
        title:
          { type: string, example: "Status API intermittently unavailable" }
        description:
          { type: string, description: "Customer report (symptom only)." }
        priority: { type: string, example: high }
        status: { $ref: "#/components/schemas/TicketStatus" }
        customer_id: { type: integer, format: int64, example: 5001 }
        customer_name: { type: string, example: "Nordlicht Logistik GmbH" }
        tags:
          type: array
          items: { type: string }
        sla_due_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time, nullable: true }
      required:
        [id, title, description, priority, status, customer_id, customer_name]

    SystemInfo:
      type: object
      properties:
        ip: { type: string, example: "10.0.0.5" }
        port: { type: integer, example: 22 }
        username: { type: string, example: azureuser }
        os: { type: string, example: "Ubuntu 22.04 LTS" }
        notes: { type: string }
      required: [ip, port, username, os]

    CustomerSystem:
      type: object
      properties:
        ticket_id: { type: integer, format: int64 }
        customer_id: { type: integer, format: int64 }
        system: { $ref: "#/components/schemas/SystemInfo" }
      required: [ticket_id, customer_id, system]

    Customer:
      type: object
      properties:
        id: { type: integer, format: int64, example: 5001 }
        company_name: { type: string, example: "Nordlicht Logistik GmbH" }
        firstname: { type: string }
        lastname: { type: string }
        system: { $ref: "#/components/schemas/SystemInfo" }
      required: [id, company_name, firstname, lastname, system]

    StatusUpdate:
      type: object
      properties:
        status: { $ref: "#/components/schemas/TicketStatus" }
      required: [status]

    ActivityCreate:
      type: object
      properties:
        ticket_id: { type: integer, format: int64, example: 7001 }
        start_datetime: { type: string, format: date-time }
        end_datetime: { type: string, format: date-time }
        description: { type: string }
        summary:
          {
            type: string,
            description: "One-sentence summary of what was restored.",
          }
        root_cause:
          {
            type: string,
            description: "Technical root cause, not the symptom.",
          }
        actions_taken:
          { type: string, description: "Diagnosis and fix steps, in order." }
        commands_summary:
          { type: string, description: "Relevant commands; no secrets." }
        validation_result:
          { type: string, description: "Concrete proof the service works." }
      required: [ticket_id, start_datetime, end_datetime]

    Activity:
      type: object
      properties:
        id: { type: integer, format: int64 }
        team_id: { type: integer }
        team_name: { type: string }
        employee_id: { type: integer, format: int64 }
        ticket_id: { type: integer, format: int64 }
        start_datetime: { type: string, format: date-time }
        end_datetime: { type: string, format: date-time }
        description: { type: string }
        summary: { type: string }
        root_cause: { type: string }
        actions_taken: { type: string }
        commands_summary: { type: string }
        validation_result: { type: string }
        created_at: { type: string, format: date-time, nullable: true }
      required:
        [
          id,
          team_id,
          team_name,
          employee_id,
          ticket_id,
          start_datetime,
          end_datetime,
          description,
        ]

    SimpleMessage:
      type: object
      properties:
        message: { type: string }
        detail: { type: object, nullable: true, additionalProperties: true }
      required: [message]

    Error:
      type: object
      properties:
        detail: { type: string }

    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items: { type: string }
              msg: { type: string }
              type: { type: string }